Break out the `login` and `logout` commands
Continuing with the clean-up of `pkg/cmd/pulumi`, this factors out the
`login` and `logout` commands into an `auth` subpackage.
